Add files via upload
Final report submitted for the internship at VW Wolfsburg, DE between 11/2022 - 05/2023. Work mostly focuses on the slanted edge method to measure the spatial frequency response of imaging systems (cameras in ADAS) and uncertainty quantification and propagation analyses in the SFR algorithm, to compute MTF as an automotive image quality metric.